requirements.txt
# Core Flask dependencies
flask>=3.1.2
flask-login>=0.6.3
flask-migrate>=4.1.0
flask-sqlalchemy>=3.1.1
flask-wtf>=1.2.2
flask-socketio>=5.5.1
flask-cors>=6.0.1
werkzeug>=3.1.3

# Database
psycopg2-binary>=2.9.10
SQLAlchemy>=2.0.20

# Image processing and barcode generation
pillow>=11.3.0
python-barcode>=0.15.1
qrcode[pil]>=7.4.2

# Payment gateways
stripe>=12.5.1
midtransclient>=1.4.2

# PDF and Excel generation
reportlab>=4.4.4
openpyxl>=3.1.5

# Environment and validation
python-dotenv>=1.1.1
email-validator>=2.3.0

# Django chat service dependencies
django>=5.2.6
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
django-extensions>=3.2.3
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0
dj-database-url>=2.1.0
asgiref>=3.7.2

# JWT for authentication
pyjwt>=2.8.0
cryptography>=41.0.8

# HTTP requests for service communication
requests>=2.31.0

# Redis (optional, fallback to in-memory for chat)
redis>=5.0.1

# Additional utilities
python-dateutil>=2.8.2
python-slugify
pytz>=2023.3

# System monitoring and process management
psutil>=5.9.0

# WebSocket support
websockets>=12.0

# File handling
python-magic>=0.4.27

# Better error handling and logging
colorlog>=6.7.0
